+++ src/NadekoBot/Common/Attributes/OwnerOnlyAttribute.cs
@@ -12,7 +12,7 @@ namespace NadekoBot.Common.Attributes
     {
         public override Task<PreconditionResult> CheckPermissionsAsync(ICommandContext context, CommandInfo executingCommand, IServiceProvider services)
         {
-            var creds = services.GetRequiredService<BotCredsProvider>().GetCreds();
+            var creds = services.GetRequiredService<IBotCredsProvider>().GetCreds();
 
             return Task.FromResult((creds.IsOwner(context.User) || context.Client.CurrentUser.Id == context.User.Id ? PreconditionResult.FromSuccess() : PreconditionResult.FromError("Not owner")));
         }